fix: 자판기 클리스를 컨트롤러화 해라
